Refactorering
