+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -59,7 +59,6 @@ namespace B2B.Transactions.Infrastructure.Configuration
             services.AddScoped<IMarketActorAuthenticator, MarketActorAuthenticator>();
             services.AddScoped<IUnitOfWork, UnitOfWork>();
             services.AddScoped<IOutgoingMessageStore, OutgoingMessageStore>();
-            services.AddScoped<IDataAvailableNotificationPublisher, DataAvailableNotificationPublisher>();
             services.AddScoped<IMessageFactory<IDocument>, AcceptMessageFactory>();
             services.AddLogging();
             AddXmlSchema(services);
